Remove callback function approach to passing
classes to @OpenAPI Decorator. Passing in a callback
function was only to be consistent with how association
decorators worked in models, where the callback functions
prevented the problem of circular class references during
file loading. That approach recently broke when one of our
projects reached a certain level of complexity, so we've
changed models to referencing one another and serializers
by string.

+++ test-app/db/migrations/1683993523103-create-pets.ts
@@ -15,7 +15,7 @@ export async function up(db: Kysely<any>): Promise<void> {
     .addColumn('collar_count_numeric', 'numeric')
     .addColumn('required_collar_count', 'bigint', col => col.notNull().defaultTo(0))
     .addColumn('required_collar_count_int', 'integer', col => col.notNull().defaultTo(0))
-    .addColumn('required_collar_count_numeric', 'integer', col => col.notNull().defaultTo(0))
+    .addColumn('required_collar_count_numeric', 'numeric', col => col.notNull().defaultTo(0))
     .addColumn('likes_walks', 'boolean')
     .addColumn('likes_treats', 'boolean', col => col.notNull().defaultTo(true))
     .addColumn('species', sql`species_types_enum`)
